Great contribution from 7-128
There are 3 apps:
DarkTimer
DarkDice
DarkKeyboard
All a result of interactions with our mutual colleague.

All of the source is included.
All of the apps have executable Jar files.
DarkKeyboard also includes an EXE wrapper.
I did not include build scripts because mine for these are specific to
an archaic environment.

7-128 has released all of the included code to the public domain.
No license at all.

The code could be useful in several ways:

1. The SAPI interface is a cheap and easy way to add self-voicing to a
Java app.
2. Ditto the sounds interface.
3. Ditto the use of a daemon on a separate thread.
4. The example apps show how one might link these solutions into apps.
5. The apps themselves are fairly good examples of how to write
quick-and-dirty java apps that don't require a lot of configuration or
externalities.
6. The included VBS script could be used by an app written in any
language.